pyproject.toml
[tool.poetry]
name = "llm_and_vector_database"
version = "0.1.0"
description = ""
authors = ["pedro lins <joao.pedrolins20@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.9.10, <3.11"
Unidecode = "^1.3.8"
python-dotenv = "^1.0.1"
pandas = "^2.2.2"
qdrant-client = "^1.9.1"
openai = "^1.34.0"
langchain-openai = "^0.1.8"
langchain = "^0.2.5"
langchain-community = "^0.2.5"
langchain-qdrant = "^0.1.1"
Flask = "^3.0.3"
gunicorn = "^22.0.0"
PyPDF2 = "^3.0.1"
streamlit = "^1.36.0"
streamlit-chat = "^0.1.1"

[tool.poetry.dev-dependencies]
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
